+++ service/src/main/kotlin/io/provenance/explorer/service/pricing/fetchers/HistoricalPriceFetcher.kt
@@ -4,5 +4,7 @@ import io.provenance.explorer.domain.models.HistoricalPrice
 import org.joda.time.DateTime
 
 interface HistoricalPriceFetcher {
+
+    fun getSource(): String
     fun fetchHistoricalPrice(fromDate: DateTime?): List<HistoricalPrice>
 }